began complete refactorization of LinkSameViewController and Board, sufficient to allow game to work at cold launch
The big idea here is that LinkSameViewController has always known too much about the Board and its view.
So now, instead of handing Board a view already in the interface, it gets the view from Board and _puts_ it in the interface.
Thus, Board is always responsible for creating a view; what it needs to know is the frame and no more.
This also means that Board is responsible for creating the hint view (createPathView).